InfoAccount como os stores querem na main bitches

+++ src/Netfreaks/Product/AbstractProductClass.java
@@ -3,6 +3,12 @@ package Netfreaks.Product;
 import java.util.HashMap;
 import java.util.Map;
 
+/**
+ *
+ * @author Bernardo Borda d'Agua    53648
+ * @author Tiago Guerreiro          53649
+ *
+ */
 abstract class AbstractProductClass implements Product{
 
     private String title;
@@ -12,7 +18,7 @@ abstract class AbstractProductClass implements Product{
     private String[] cast;
     private String masterName;
     private Map<String, Integer> rates;
-    private double averageRating;
+    private float averageRating;
 
     AbstractProductClass(String title, int yearOfRelease, String PEGI, String genre, String[] cast) {
         this.title = title;
@@ -69,7 +75,7 @@ abstract class AbstractProductClass implements Product{
 
     @Override
     public float getAverageRating(){
-        return (float)averageRating;
+        return averageRating;
     }
 
     @Override
